# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Outreach Service API
  version: "1.0.0"
  description: API per gestione outreach email aziende
servers:
  - url: https://outreach.example.com
    description: Produzione
  - url: http://localhost:3002
    description: Sviluppo
paths:
...
  /api/admin/settings/{key}:
    put:
      summary: Aggiorna impostazione (admin only)
      parameters:
        - { name: key, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                value: { type: string }
      responses:
        "200": { description: Aggiornata, SMTP riconfigurato se outreach_smtp_* }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
security:
  - cookieAuth: []